refactor: split interpreter into env and eval modules
