+++ src/compiler/intrinsic/net.rs
@@ -1,4 +1,4 @@
-use crate::compiler::interp::value::Value;
+use crate::compiler::interp::Value;
 use crate::runtime::ffi::net;
 use std::rc::Rc;
